chore: export SessionDataStore and LogoutToken types
