+++ src/server/session/stateful-session-store.ts
@@ -1,9 +1,6 @@
-import { SessionData } from "../../types"
+import { SessionData, SessionDataStore } from "../../types"
 import * as cookies from "../cookies"
-import {
-  AbstractSessionStore,
-  SessionDataStore,
-} from "./abstract-session-store"
+import { AbstractSessionStore } from "./abstract-session-store"
 
 // the value of the stateful session cookie containing a unique session ID to identify
 // the current session
